FluxBank/requirements.txt
Flask==2.0.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
Werkzeug==2.0.1
email-validator==1.1.3
python-dotenv==0.19.0
Flask-Mail==0.9.1
PyJWT==2.1.0
bcrypt==3.2.0
mysql-connector-python==8.0.33
python-dotenv==0.19.0 